requirements.txt
Django==4.1.1
djangorestframework==3.14.0
drf_yasg==1.21.4
